+++ crates/cubecl-ir/src/atomic.rs
@@ -1,11 +1,13 @@
 use std::fmt::Display;
 
+use cubecl_macros::OperationCore;
 use type_hash::TypeHash;
 
 use crate::{BinaryOperator, CompareAndSwapOperator, Operation, UnaryOperator};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, Clone, TypeHash, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, TypeHash, PartialEq, Eq, Hash, OperationCore)]
+#[operation(opcode_name = AtomicOpCode)]
 pub enum AtomicOp {
     Load(UnaryOperator),
     Store(UnaryOperator),
