+++ crates/cubecl-ir/src/cmma.rs
@@ -1,8 +1,9 @@
 use std::fmt::Display;
 
+use smallvec::SmallVec;
 use type_hash::TypeHash;
 
-use crate::Operation;
+use crate::{Operation, OperationCode, OperationCore};
 
 use super::{Elem, Variable};
 
@@ -38,7 +39,8 @@ pub struct Matrix {
 
 /// Cooperative Matrix-Multiply and Accumulate Instruction.
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, Clone, TypeHash, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, TypeHash, PartialEq, Eq, Hash, OperationCode)]
+#[operation(opcode_name = CmmaOpCode)]
 #[allow(missing_docs)]
 pub enum CoopMma {
     /// Fill the matrix with the value.
@@ -65,6 +67,30 @@ pub enum CoopMma {
     },
     /// Cast a fragment to another type.
     Cast { input: Variable },
+}
+
+impl OperationCore for CoopMma {
+    type OpCode = CmmaOpCode;
+
+    fn op_code(&self) -> Self::OpCode {
+        self.__match_opcode()
+    }
+
+    fn args(&self) -> Option<smallvec::SmallVec<[Variable; 4]>> {
+        match self {
+            CoopMma::Fill { value } => Some(SmallVec::from_slice(&[*value])),
+            CoopMma::Load { .. } | CoopMma::Execute { .. } | CoopMma::Store { .. } => None,
+            CoopMma::Cast { input } => Some(SmallVec::from_slice(&[*input])),
+        }
+    }
+
+    fn from_code_and_args(op_code: Self::OpCode, args: &[Variable]) -> Option<Self> {
+        match op_code {
+            CmmaOpCode::Fill => Some(CoopMma::Fill { value: args[0] }),
+            CmmaOpCode::Load | CmmaOpCode::Execute | CmmaOpCode::Store => None,
+            CmmaOpCode::Cast => Some(CoopMma::Cast { input: args[0] }),
+        }
+    }
 }
 
 impl Display for CoopMma {
