--- crates/cubecl-ir/src/element.rs
@@ -1,1 +1,0 @@
-
