+++ crates/cubecl-ir/src/lib.rs
@@ -3,11 +3,12 @@ mod atomic;
 mod base;
 mod branch;
 mod cmma;
-mod element;
 mod item;
 mod macros;
+mod metadata;
 mod non_semantic;
 mod operation;
+mod operator;
 mod pipeline;
 mod plane;
 mod processing;
@@ -21,10 +22,14 @@ pub use base::*;
 pub use branch::*;
 pub use cmma::*;
 pub use item::*;
+pub use metadata::*;
 pub use non_semantic::*;
 pub use operation::*;
+pub use operator::*;
 pub use pipeline::*;
 pub use plane::*;
 pub use scope::*;
 pub use synchronization::*;
 pub use variable::*;
+
+pub(crate) use cubecl_macros::{OperationArgs, OperationCode, OperationCore};
