+++ crates/cubecl-ir/src/scope.rs
@@ -289,7 +289,7 @@ impl Scope {
         let mut operations = Vec::new();
 
         for (local, scalar) in self.reads_scalar.drain(..) {
-            operations.push(Instruction::new(Operation::Copy(scalar), local));
+            operations.push(Instruction::new(Operation::Copy(scalar.into()), local));
             variables.push(local);
         }
