+++ crates/cubecl-macros/src/parse/operation.rs
@@ -6,8 +6,20 @@ use darling::{
 use syn::{Generics, Ident, Type, Visibility};
 
 #[derive(FromDeriveInput)]
-#[darling(supports(enum_newtype), attributes(operation))]
+#[darling(supports(enum_newtype, enum_unit, enum_named), attributes(operation))]
 pub struct Operation {
+    pub ident: Ident,
+    pub vis: Visibility,
+    pub generics: Generics,
+    pub data: Data<OperationVariant, ()>,
+
+    pub opcode_name: Ident,
+    pub with_children: Flag,
+}
+
+#[derive(FromDeriveInput)]
+#[darling(supports(enum_any), attributes(operation))]
+pub struct OpCode {
     pub ident: Ident,
     pub vis: Visibility,
     pub generics: Generics,
@@ -25,5 +37,6 @@ pub struct OperationVariant {
 
 #[derive(FromField)]
 pub struct OperationField {
+    pub ident: Option<Ident>,
     pub ty: Type,
 }
